% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fermicatsR2.R
\docType{data}
\name{FGL4DR2}
\alias{FGL4DR2}
\title{4FGLDR2 Catalog (Fourth Fermi Large Area Telescope Source Catalog, Second Data Release)}
\format{
A data frame with 142 variables on 5788 gamma-ray sources.
}
\source{
\url{http://fermi.gsfc.nasa.gov/ssc/data/access/lat/10yr_catalog/}
}
\usage{
FGL4DR2
}
\description{
Fermi Large Area Telescope Source Catalog Second Data Release (4FGLDR2).
Ballet, J. et al., arXiv, 2005.11208v1, (2020).
}
\details{
Initial Release: 22 May 2020
Latest Release: gll_psc_v23.fit (22 May 2020)
}
\section{Fields}{

\describe{
\item{Source_Name}{4FGL JHHMM.m+DDMM\link{c/e}, constructed according to IAU Specifications for Nomenclature;
m is decimal minutes of R.A.; in the name R.A. and decl. are truncated at 0.1 decimal minutes and 1' ,
respectively; 'c' indicates that based on the region of the sky the source is considered to be potentially confused with
Galactic diffuse emission; e indicates a source that was modeled as spatially extended (see Table 12 of 4FGL paper)}
\item{RAJ2000}{Right Ascension, J2000, deg}
\item{DEJ2000}{Declination, J2000, deg}
\item{GLON}{Galactic longitude, deg}
\item{GLAT}{Galactic latitude, deg}
\item{Conf_68_SemiMajor}{Long radius of error ellipse at 68\\% confidence, deg}
\item{Conf_68_SemiMinor}{Short radius of error ellipse at 68\\% confidence, deg}
\item{Conf_68_PosAng}{Position angle of the 68\\% confidence ellipse, deg east of north}
\item{Conf_95_SemiMajor}{Long radius of error ellipse at 95\\% confidence, deg}
\item{Conf_95_SemiMinor}{Short radius of error ellipse at 95\\% confidence, deg}
\item{Conf_95_PosAng}{Position angle of (eastward) of the long axis from celestial North, deg}
\item{ROI_num}{ROI number (cross-reference to ROIs extension)}
\item{Extended_Source_Name}{Cross-reference to the ExtendedSources extension}
\item{Signif_Avg}{Source significance (in sigma units) derived from likelihood TS for 100 MeV\342\200\223300 GeV analysis, one decimal place}
\item{Pivot_Energy}{Energy at which error on differential flux is minimal, MeV}
\item{Flux1000}{Integral flux from 1 to 100 GeV, cm^{-2} s^{-1}}
\item{Unc_Flux1000}{1 sigma error on integral flux from 1 to 100 GeV, cm^{-2} s^{-1}}
\item{Energy_Flux100}{Energy flux from 100 MeV to 100 GeV obtained by spectral fitting, erg cm^{-2} s^{-1}}
\item{Unc_Energy_Flux100}{1 sigma error on energy flux from 100 MeV to 100 GeV, erg cm^{-2} s^{-1}}
\item{SpectrumType}{Spectral type (PowerLaw, LogParabola, PLExpCutoff, PLSuperExpCutoff)}
\item{PL_Flux_Density}{Differential flux at Pivot_Energy in PowerLaw fit, cm^{-2} MeV^{-1} s^{-1}}
\item{Unc_PL_Flux_Density}{1 sigma error on PL_Flux_Density, cm^{-2} MeV^{-1} s^{-1}}
\item{PL_Index}{Photon index when fitting with PowerLaw}
\item{Unc_PL_Index}{1 sigma error on PL_Index}
\item{LP_Flux_Density}{Differential flux at Pivot_Energy in LogParabola fit, cm^{-2} MeV^{-1} s^{-1}}
\item{Unc_LP_Flux_Density}{1 sigma error on LP_Flux_Density, cm^{-2} MeV^{-1} s^{-1}}
\item{LP_Index}{Photon index at Pivot_Energy (alpha of Eq. 2) when fitting with LogParabola}
\item{Unc_LP_Index}{1 sigma error on LP_Index}
\item{LP_beta}{Curvature parameter (beta of Eq. 2) when fitting with LogParabola}
\item{Unc_LP_beta}{1 sigma error on LP_beta}
\item{LP_SigCurv}{Significance (in sigma units) of the fit improvement between PowerLaw and LogParabola. A value greater than 4 indicates significant curvature}
\item{PLEC_Flux_Density}{Differential flux at Pivot_Energy in PLSuperExpCutoff fit, cm^{-2} MeV^{-1} s^{-1}}
\item{Unc_PLEC_Flux_Density}{1 sigma error on PLEC_Flux_Density, cm^{-2} MeV^{-1} s^{-1}}
\item{PLEC_Index}{Low-energy photon index (gamma of Eq. 4) when fitting with PLSuperExpCutoff}
\item{Unc_PLEC_Index}{1 sigma error on PLEC_Index}
\item{PLEC_Expfactor}{Exponential factor (a of Eq. 4) when fitting with PLSuperExpCutoff}
\item{Unc_PLEC_Expfactor}{1 sigma error on PLEC_Expfactor}
\item{PLEC_Exp_Index}{Exponential index (b of Eq. 4) when fitting with PLSuperExpCutoff}
\item{Unc_PLEC_Exp_Index}{1 sigma error on PLEC_Exp_Index}
\item{PLEC_SigCurv}{Same as LP_SigCurv for PLSuperExpCutoff model}
\item{Npred}{Predicted number of events in the model}
\item{Flux_Band}{Integral photon flux in each spectral band, cm^{-2} s^{-1}}
\item{Unc_Flux_Band}{1 sigma lower and upper error on Flux_Band, cm^{-2} s^{-1}}
\item{nuFnu_Band}{Spectral energy distribution over each spectral band, erg cm^{-2} s^{-1}}
\item{Sqrt_TS_Band}{Square root of the Test Statistic in each spectral band}
\item{Variability_Index}{Sum of 2*log(Likelihood) difference between the flux fitted in each time interval and the average flux over the full catalog interval; a value greater than 18.48 over 12 intervals indicates <1\\% chance of being a steady source}
\item{Frac_Variability}{Fractional variability computed from the fluxes in each year}
\item{Unc_Frac_Variability}{1 sigma error on fractional variability}
\item{Signif_Peak}{Source significance in peak interval in sigma units}
\item{Flux_Peak}{Peak integral photon flux from 100 MeV to 100 GeV, cm^{-2} s^{-1}}
\item{Unc_Flux_Peak}{1 sigma error on peak integral photon flux, cm^{-2} s^{-1}}
\item{Time_Peak}{Time of center of interval in which peak flux was measured, s (MET)}
\item{Peak_Interval}{Length of interval in which peak flux was measured, s}
\item{Flux_History}{Integral photon flux from 100 MeV to 100 GeV in each year (best fit from likelihood analysis with spectral shape fixed to that obtained over full interval), cm^{-2} s^{-1}}
\item{Unc_Flux_History}{1 sigma lower and upper error on integral photon flux in each year, cm^{-2} s^{-1}}
\item{Sqrt_TS_History}{Square root of the Test Statistic in each year}
\item{ASSOC_FGL}{Most recent correspondence to previous FGL source catalogs, if any}
\item{ASSOC_FHL}{Most recent correspondence to previous FHL source catalogs, if any}
\item{ASSOC_GAM1}{Name of likely corresponding 2AGL source, if any}
\item{ASSOC_GAM2}{Name of likely corresponding 3EG source, if any}
\item{ASSOC_GAM3}{Name of likely corresponding EGR source, if any}
\item{TEVCAT_FLAG}{P if positional association with non-extended source in TeVCat, E if associated with an extended source in TeVCat, N if no TeV association}
\item{ASSOC_TEV}{Name of likely corresponding TeV source from TeVCat, if any}
\item{CLASS1}{Class designation for associated source; see Table 7}
\item{CLASS2}{Class designation for low-confidence association}
\item{ASSOC1}{Name of identified or likely associated source}
\item{ASSOC2}{Name of low-confidence association or of enclosing extended source}
\item{ASSOC_PROB_BAY}{Probability of association according to the Bayesian method}
\item{ASSOC_PROB_LR}{Probability of association according to the Likelihood Ratio method}
\item{RA_Counterpart}{Right Ascension of the counterpart ASSOC1, deg}
\item{DEC_Counterpart}{Declination of the counterpart ASSOC1, deg}
\item{Unc_Counterpart}{95\\% precision of the counterpart localization, deg}
\item{Flags}{Source flags (binary coding as in Table 5)}
}
}

\keyword{datasets}
