% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fermicatsR2.R
\docType{data}
\name{FHL3}
\alias{FHL3}
\title{3FHL Catalog (Third Catalog of Hard Fermi-LAT Sources)}
\format{
A data frame with 65 variables on 1556 sources.
}
\source{
\url{http://fermi.gsfc.nasa.gov/ssc/data/access/lat/3FHL/}
}
\usage{
FHL3
}
\description{
The Third Catalog of Hard Fermi-LAT Sources (3FHL).
Ajello, M. et al., The Astrophysical Journal Supplement Series, 232, 18 (2017).
FITS Filename: gll_psch_v13.fit, released 19 Jul 2017.
}
\section{Fields}{

\describe{
\item{Source_Name}{3FHL JHHMM.m+DDMM, constructed according to IAU Specifications for Nomenclature;
m is decimal minutes of R.A.; in the name R.A. and decl. are truncated at 0.1 decimal minutes and 1' , respectively. A Source_Name ending with "e" indicates an extended source.}
\item{RAJ2000}{Right Ascension, J2000}
\item{DEJ2000}{Declination, J2000}
\item{GLON}{Galactic longitude, deg.}
\item{GLAT}{Galactic latitude, deg.}
\item{Conf_95_SemiMajor}{Error radius at 95\\% confidence, deg.}
\item{Conf_95_SemiMinor}{=Conf_95_SemiMajor in 3FHL, deg.}
\item{Conf_95_PosAng}{NULL in 3FHL (error circles), deg.}
\item{ROI_num}{ROI number (cross-reference to ROIs extension)}
\item{Signif_Avg}{Source significance in sigma units over the 10 GeV to 2 TeV band}
\item{Pivot_Energy}{Energy at which error on differential flux is minimal, GeV}
\item{Flux_Density}{Differential flux at Pivot_Energy, cm^{-2} GeV^{-1} s^{-1}}
\item{Unc_Flux_Density}{1 sigma error on differential flux at Pivot_Energy, cm^{-2} GeV^{-1} s^{-1}}
\item{Flux}{Integral photon flux from 10 GeV to 1 TeV obtained by spectral fitting, cm^{-2} s^{-1}}
\item{Unc_Flux}{1 sigma error on integral photon flux from 10 GeV to 1 TeV, cm^{-2} s^{-1}}
\item{Energy_Flux}{Energy flux from 10 GeV to 1 TeV obtained by spectral fitting, erg cm^{-2} s^{-1}}
\item{Unc_Energy_Flux}{1 sigma error on energy flux from 10 GeV to 1 TeV, erg cm^{-2} s^{-1}}
\item{Signif_Curve}{Significance (in sigma units) of the fit improvement between power-law and LogParabola. A value greater than 3 indicates significant curvature}
\item{SpectrumType}{Spectral type (PowerLaw or LogParabola)}
\item{Spectral_Index}{Best-fit photon number index at Pivot_Energy when fitting with LogParabola}
\item{Unc_Spectral_Index}{1 sigma error on Spectral_Index}
\item{beta}{Curvature parameter beta when fitting with LogParabola}
\item{Unc_beta}{1 sigma error on beta}
\item{PowerLaw_Index}{Best-fit photon number index when fitting with power law}
\item{Unc_PowerLaw_Index}{1 sigma error on PowerLaw_Index}
\item{Flux_Band}{Integral photon flux in each spectral band, cm^{-2} s^{-1}}
\item{Unc_Flux_Band}{1 sigma lower and upper error on Flux_Band, cm^{-2} s^{-1}}
\item{nuFnu}{Spectral energy distribution over each spectral band, erg cm^{-2} s^{-1}}
\item{Sqrt_TS_Band}{Square root of the Test Statistic in each spectral band}
\item{Npred}{Predicted number of events in the model}
\item{HEP_Energy}{Highest energy among events probably coming from the source, GeV}
\item{HEP_Prob}{Probability of that event to come from the source}
\item{Variability_BayesBlocks}{Number of Bayesian blocks from variability analysis; 1 if not variable, −1 if could not be tested}
\item{Extended_Source_Name}{Cross-reference to the ExtendedSources extension}
\item{ASSOC_GAM}{Correspondence to previous gamma-ray source catalog}
\item{TEVCAT_FLAG}{P if positional association with non-extended source in TeVCat, E if associated with an extended source in TeVCat, N if no TeV association, C if TeV source candidate as defined in Section 3.4}
\item{ASSOC_TEV}{Name of likely corresponding TeV source from TeVCat, if any}
\item{CLASS}{Class designation for associated source; see Table 2}
\item{ASSOC1}{Name of identified or likely associated source}
\item{ASSOC2}{Alternate name or indicates whether the source is inside an extended source}
\item{ASSOC_PROB_BAY}{Probability of association according to the Bayesian method}
\item{ASSOC_PROB_LR}{Probability of association according to the Likelihood-Ratio method}
\item{Redshift}{Redshift of counterpart, if known}
\item{NuPeak_obs}{Frequency of the synchrotron peak of counterpart, if known, Hz}
}
}

\keyword{datasets}
