% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fermicatsR2.R
\docType{data}
\name{LAC4_LO}
\alias{LAC4_LO}
\title{4LAC_LO (Fourth Catalog of Active Galactic Nuclei Detected by the Fermi Large Area Telescope - Low Galactic Latitude)}
\format{
A data frame with 38 variables on 344 gamma-ray sources.
}
\source{
\url{https://ui.adsabs.harvard.edu/abs/2020ApJ...892..105A/abstract}
}
\usage{
LAC4_LO
}
\description{
Fourth Catalog of Active Galactic Nuclei (4LAC).
Ajello, M. et al., The Astrophysical Journal, 892, 105 (2020).
}
\details{
Low Galactic Latitude (|GLAT| < 10 deg.) Sources.
}
\section{Fields}{

\describe{
\item{Source_Name}{Source name 4FGL JHHMM.m+DDMMa}
\item{RAJ2000}{R.A., deg}
\item{DEJ2000}{Decl., deg}
\item{GLON}{Galactic Longitude, deg}
\item{GLAT}{Galactic Latitude, deg}
\item{Signif_Avg}{Source significance in sigma units over the 50 MeV–1 TeV band}
\item{Flux1000}{Integral photon flux from 1 to 100 GeV, cm^{−2} s^{−1}}
\item{Unc_Flux1000}{1 sigma error on integral photon flux from 1 to 100 GeV, cm^{−2} s^{−1}}
\item{Energy_Flux100}{Energy flux from 100 MeV to 100 GeV obtained by spectral fitting, erg cm^{−2} s^{−1}}
\item{Unc_Energy_Flux100}{1 sigma error on energy flux from 100 MeV to 100 GeV, erg cm^{−2} s^{−1}}
\item{SpectrumType}{Spectral type in the global model (PowerLaw, LogParabola, PLSuperExpCutoff)}
\item{PL_Index}{Photon index when fitting with PowerLaw}
\item{Unc_PL_Index}{1 sigma error on PL_Index}
\item{Pivot_Energy}{Pivot Energy, MeV}
\item{LP_Index}{Photon index at Pivot_Energy (alpha) when fitting with LogParabola}
\item{Unc_LP_Index}{1 sigma error on LP_Index}
\item{LP_beta}{Curvature parameter (beta) when fitting with LogParabola}
\item{Unc_LP_beta}{1 sigma error on LP_beta}
\item{Flags}{Analysis flags}
\item{CLASS}{Class designation for associated source}
\item{ASSOC1}{Name of identified or likely associated source}
\item{ASSOC_PROB_BAY}{Probability of association according to the Bayesian method}
\item{ASSOC_PROB_LR}{Probability of association according to the Likelihood Ratio method}
\item{Counterpart_Catalog}{Counterpart catalog driving the association}
\item{RA_Counterpart}{R.A. of the counterpart ASSOC1, deg}
\item{DEC_Counterpart}{Decl. of the counterpart ASSOC1, deg}
\item{Unc_Counterpart}{95\\% precision of the counterpart localization, deg}
\item{VLBI_Counterpart}{Name of the VLBI counterpart}
\item{Gaia_Counterpart}{Name of the Gaia counterpart established via the VLBI position}
\item{Gaia_G_Magnitude}{Gaia Magnitude at the G band}
\item{Redshift}{Redshift}
\item{SED_class}{SED-based class}
\item{nu_syn}{Synchrotron-peak frequency in observer frame, Hz}
\item{nuFnu_syn}{Spectral energy distribution at synchrotron-peak frequency, erg cm^{−2} s^{−1}}
\item{Variability_Index}{Variability index}
\item{Frac_Variability}{Fractional variability}
\item{Unc_Frac_Variability}{1 sigma error on fractional variability}
\item{Highest_energy}{Energy of the highest-energy ULTRACLEANVETO photon with association probability P>0.95, GeV}
}
}

\keyword{datasets}
